pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "abstractassistant"
version = "1.0.0"
description = "A sleek macOS system tray application providing instant access to LLMs"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "AbstractAssistant Team"}
]
keywords = ["ai", "llm", "macos", "system-tray", "assistant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Desktop Environment",
]

dependencies = [
    "abstractcore[all]>=2.4.0",
    "pystray>=0.19.4",
    "Pillow>=10.0.0",
    "customtkinter>=5.2.0",
    "markdown>=3.5.0",
    "pyperclip>=1.8.2",
    "plyer>=2.1.0",
    "tomli>=2.0.0; python_version<'3.11'",
    "tomli-w>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[project.scripts]
assistant = "abstractassistant.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/abstractassistant"
Repository = "https://github.com/yourusername/abstractassistant"
Issues = "https://github.com/yourusername/abstractassistant/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["abstractassistant*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
